# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Personas",
        "description": "Mostando la Lista de URI's de mi API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://PruebaTecnicaFourgen.test"
        }
    ],
    "paths": {
...
        "/api/pet": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Obtener todas las mascotas",
                "operationId": "5eb031c45f88e3872abd535c9f33cfb9",
                "responses": {
                    "200": {
                        "description": "Lista de todas las mascotas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pet"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las mascotas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error fetching pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Crear una nueva mascota",
                "operationId": "6e8419c722e12d6f56f296a172eb086f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mascota creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear la mascota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error creating pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pet/{id}": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Obtener una mascota por ID",
                "operationId": "b89fa6a55fa523dc12d24c117cc3cdf3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la mascota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mascota no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pet not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener la mascota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error fetching pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Pets"
                ],
                "summary": "Actualizar los datos de una mascota",
                "operationId": "cb99e0397ab4f2ec7f973c546f3f1e14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mascota actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar la mascota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error updating pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pets"
                ],
                "summary": "Eliminar una mascota",
                "operationId": "9d8e480efe7aad7cb48d42cf3ffa39d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mascota eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pet deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar la mascota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error deleting pet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreatePetRequest": {
                "title": "Crear mascota",
                "required": [
                    "name",
                    "species",
                    "person_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Toby"
                    },
                    "species": {
                        "type": "string",
                        "example": "Gato"
                    },
                    "person_id": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            },
...
            "UpdatePetRequest": {
                "title": "Actualizar mascota",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Toby"
                    },
                    "species": {
                        "type": "string",
                        "example": "Gato"
                    },
                    "person_id": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            },
...
            "Pet": {
                "title": "Pet",
                "description": "Modelo de mascota",
                "required": [
                    "name",
                    "species",
                    "person_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Firulais"
                    },
                    "species": {
                        "type": "string",
                        "example": "Perro"
                    },
                    "breed": {
                        "type": "string",
                        "example": "Labrador"
                    },
                    "age": {
                        "type": "integer",
                        "example": 3
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/images/firulais.jpg"
                    },
                    "person_id": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            }
        }
    },
...
}